Flexibly show issues or pull requests in Show
Show now displays issues or pull requests (it did that before, but as
pull requests have a little more information, a difference needs to be
made).
The requested issue number gets used in a parallel dual call to the
API: Both the issue and the pull API's are hit. If a pull request was
successfully retrieved, it is shown instead of the plain issue view.

The user won't notice the difference when he makes the call, but will
rejoice upon the better PR output in his terminal.