Fix minor formatting problem in format_diff
Something along the following imposed problems:

+
+ def

The second line would not be highlighted properly, but only when the
pager is active.  The [^\+] part of the regex always matched the newline
following the '+' on the first line, the regex therefore matched multiple
lines and took the '+ def' with it - the reset color escape sequence
was put in right after the second line.

This is not a problem for a regular ruby puts, which flushes $stdout at
once, the highlighting can span multiple lines that way.
However with a pager like less every line gets printed on the screen
naturally - the second line had no escape sequence, it was white.

Everything's fine now.