Add --no-labels option to `ghi-list`
There have been a few requests for an option to prevent printing labels
when listing issues. If the list of issues involves a lot of labels, it
leads to truncation of issue titles. To make it easier to visually parse
lists like this, I've added a `--no-labels` option to `ghi-list` to
skip printing labels in the pager. Closes #259.

Signed-off-by: David Celis <me@davidcel.is>