Use application/vnd.github.v3+json
The `application/vnd.github.beta+json` media type has been
[deprecated](https://developer.github.com/v3/versions/) for quite some
time. It's recommended that the actual API version, v3, be used instead.

This had only a minor impact to GHI: in v3, the `"pull_request"` key of
an issue response is omitted for issues that are not also pull requests.
Previously, `"pull_request"` would still be present but empty.

Signed-off-by: David Celis <me@davidcel.is>