Add [ -- [<user>/]<repo> ] suffix to warning
In #279, a user expressed confusion that `[ -- [<user>/]<repo> ]` as a
command suffix is actually global and can be used with any command. This
patch adds a helpful message to the warning "Not a GitHub repo" to
suggest either re-running the command from a GitHub repo, or using the
suffix.

Fixes #279.

Signed-off-by: David Celis <me@davidcel.is>