Add a --verbose option for listing labels
When --verbose (-v) is passed to the `ghi label -l -v` command, it
shows the hex codes for the colors of the labels instead of just using
the color on the background in the terminal.

This allows users to know what colors are already being used in case
they want to make a matching set of labels for their repository.